# ICTTaxApi/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ICTTaxApi",
    "version": "1.0"
  },
  "paths": {
    "/api/client/transactions": {
      "get": {
        "tags": [
          "ICTTransaction"
        ],
        "operationId": "GetTransactions",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TransactionDTO"
              }
            }
          }
        }
      }
    },
    "/api/client": {
      "get": {
        "tags": [
          "ICTTransaction"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TransactionDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ICTTransaction"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/client/{clientname}": {
      "get": {
        "tags": [
          "ICTTransaction"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clientname",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TransactionDTO"
              }
            }
          }
        }
      }
    },
    "/api/client/summary": {
      "get": {
        "tags": [
          "ICTTransaction"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/TransactionSummaryDTO"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "TransactionDTO": {
      "required": [
        "amount",
        "clientName",
        "transactionDate"
      ],
      "type": "object",
      "properties": {
        "transactionDate": {
          "type": "string"
        },
        "clientName": {
          "type": "string"
        },
        "description": {
          "maxLength": 120,
          "minLength": 0,
          "type": "string"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "filename": {
          "type": "string"
        },
        "uploadedDate": {
          "type": "string"
        }
      }
    },
    "TransactionSummaryDTO": {
      "type": "object",
      "properties": {
        "totalTransactions": {
          "format": "int32",
          "type": "integer"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "fileCount": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}